Move all components into a component folder
