+++ src/components/Form/Form.css
@@ -31,7 +31,7 @@
   display: flex;
   flex-direction: column;
   align-items: center;
-  background-image: url('../assets/images/chat-bubble.png');
+  background-image: url('../../assets/images/chat-bubble.png');
   background-size: 28vw;
   background-repeat: no-repeat, repeat;
   background-position: center;
