+++ src/components/Main/Main.js
@@ -3,7 +3,7 @@ import './Main.css'
 import Form from '../Form/Form'
 import Results from '../Results/Results'
 import Card from '../Card/Card'
-import isabella from '../assets/images/isabella.gif'
+import isabella from '../../assets/images/isabella.gif'
 import { Route } from 'react-router-dom'
 
 function Main() {
